printf new

+++ libft/convert/ft_lutoa_base.c
@@ -6,13 +6,13 @@
 /*   By: zraunio <zraunio@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/06 11:24:23 by zraunio           #+#    #+#             */
-/*   Updated: 2021/01/11 10:25:09 by zraunio          ###   ########.fr       */
+/*   Updated: 2021/02/26 15:02:45 by zraunio          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../libft.h"
 
-static int	ft_base(unsigned long int nb)
+static int	ft_base(unsigned long long int nb)
 {
 	if (nb >= 10)
 		return (nb - 10 + 'a');
@@ -20,11 +20,11 @@ static int	ft_base(unsigned long int nb)
 		return (nb + '0');
 }
 
-char		*ft_lutoa_base(unsigned long int nb, unsigned long int base)
+char		*ft_lutoa_base(unsigned long long int nb, unsigned long int base)
 {
-	int					i;
-	char				*res;
-	unsigned long int	tmp;
+	int						i;
+	char					*res;
+	unsigned long long int	tmp;
 
 	i = 0;
 	tmp = nb;
